+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/gecko/GeckoProvider.kt
@@ -109,6 +109,12 @@ object GeckoProvider {
             .translationsOfferPopup(context.settings().offerTranslation)
             .disableShip(FxNimbus.features.ship.value().disabled)
             .fissionEnabled(FxNimbus.features.fission.value().enabled)
+            .setSameDocumentNavigationOverridesLoadType(
+                FxNimbus.features.sameDocumentNavigationOverridesLoadType.value().enabled,
+            )
+            .setSameDocumentNavigationOverridesLoadTypeForceDisable(
+                FxNimbus.features.sameDocumentNavigationOverridesLoadType.value().forceDisableUri,
+            )
             .build()
     }
 }